+++ Code/diorama_private_files.cmake
@@ -30,6 +30,7 @@ set(FILES
     Source/Clients/Particles2D.h
     Source/Clients/SpriteCull.h
     Source/Clients/SpriteTrail.h
+    Source/Clients/SpritePalette.h
     Source/Clients/SpriteFeatureProcessor.cpp
     Source/Clients/SpriteFeatureProcessor.h
     Source/Clients/SpritePresenter.cpp